Move tab components to different folder
