+++ src/components/GettingStarted/Steps/ConfigureKubectl.js
@@ -6,7 +6,6 @@ import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
 import { bindActionCreators } from 'redux';
 import { OrganizationsRoutes } from 'shared/constants/routes';
-import { Tab, Tabs } from 'shared/Tabs';
 import * as clusterActions from 'stores/cluster/actions';
 import { getLoggedInUser } from 'stores/main/selectors';
 import styled from 'styled-components';
@@ -15,6 +14,7 @@ import Button from 'UI/Controls/Button';
 import ClusterIDLabel from 'UI/Display/Cluster/ClusterIDLabel';
 import { CodeBlock, Prompt } from 'UI/Display/Documentation/CodeBlock';
 import GettingStartedBottomNav from 'UI/Display/Documentation/GettingStartedBottomNav';
+import { Tab, Tabs } from 'UI/Display/Tabs';
 import Aside from 'UI/Layout/Aside';
 
 import ConfigureKubeCtlAlternative from './ConfigureKubectlAlternative';
